Options parsing
We are going to use the package optparse-applicative to create an
program options parser.

We would like to add the following options:

1. input directory
2. output directory
3. a flag signaling that we can replace the output folder if it
   already exists

optparse-applicative provides us with an interface to write program
arguments parsers in "applicative" style. Applicative is another type
class like Monoid, but work on higher structures such as `Maybe`, lists
and `IO`. The typeclassopedia is a good place to read about these kind
of typeclasses. Read about Functor and Applicative here:
https://wiki.haskell.org/Typeclassopedia

You might also want to learn about Kinds to understand the different
between for example `Maybe` and and `Int`:
https://github.com/soupi/haskell-study-plan/#kinds

We are defining a new data type to hold our configuration info and are
defining a parsing using the EDSL defined in the optparse-applicative
package to build our parser.

Check out the readme section of optparse-applicative in hackage for
more details:

https://hackage.haskell.org/package/optparse-applicative

This provides us with a pretty declarative interface for building
option parsers which includes helper messages, commands, usage
message, etc.