Replacing String with Text (language extensions, guards)
We mentioned before that the String type is just an alias to a list of Char.
While that makes the API pretty easy to implement and use, it often
has dire implications on program performance and memory usage.

For more information, please read this article:
https://www.alexeyshmalko.com/2015/haskell-string-types/

In this commit we will replace our usage of String with Text. And while
the changes are pretty minimal, we are using a few new features - so
lets talk about them.

(1) Qualified Imports

Up until now, whenever we imported a module that module brought to
scope all of the names defined in that module. In the case of
`Data.Text` and many other modules (Such as `Data.ByteString`, `Data.Map`,
`Data.Set` etc) there will be name conflicts with each other and
`Prelude` for many names such as `map`, `concat` and more.

These conflicts makes it easy to figure out the name of functions,
they are all named the same way, but requires us to "qualify" which
function from which module we'd like to use which name. We can do that
by importing modules as qualified we can differentiate between similar
names from different modules.

Qualify a module like this:

```
module qualified <module> as <qualifier>
```

and then use it like this: `<qualifier>.<name>`.

More about imports here: https://wiki.haskell.org/Import

(2) GHC Language extensions

The GHC compiler provides extensions to the Haskell language beyond
what's defined in the Haskell report. These extensions are somewhat
experimental and can be activated using flags or special comments.

`OverloadedStrings` is such extension, normally, the type of string
literals is `String`. Using this extension the type of string literals
is more general and is inferred from its use. So the type of string
literals could be Text, ByteString, or something else. This makes it
easier to use Text as the default String type which is what we want.

We activate it by writing a special comment at the top of the module:

`{-# language <extension-name> #-}`

There are many extensions, not all of them important, but some can be
very useful.

You can learn more about extensions in these links:

- https://typeclasses.com/extensions-intro
- https://ocharles.org.uk/pages/2014-12-01-24-days-of-ghc-extensions.html

(3) Guards syntax

With guards we can write an multi way `if` like control structure when
defining a function. We can define either a boolean expression or some
kind of pattern, and if it matches the right side will be returned.

```
myfunc <args>
  | <condition or pattern> = <expr>
  ...
  | otherwise = <expr>
```

For patterns, we use `<pattern> <- <case>` syntax which is kinda
similar to:

```
case <case> of
  | <pattern> -> ...
```

Since we cannot pattern match on Text the same way we did with
`String`s (which are linked lists). we'll be a bit more creative and
create a function for our usage and use guards instead.