Bonus: Serve blog over http (wai)
In this commit we'll add an alternative way to use our program and in
addition to converting a directory with text files to html, we will
also add a command to serve these files over http!

We'll change our program interface to include 2 commands,
`convert` and `serve`. `convert` we'll have the same arguments we used
up until now and `serve` will get a port and an input directory.

(1) app/blog_gen.hs

In app/blog_gen.hs, We'll add a new data type called `Command` to
differentiate the two commands, and we'll use the
optparse-applicative function `subparser` to merge two command parsers.

(2) src/Process.hs

We'll extract the functionality that reads the files from the
directory and creates Html docs out of them into a separate function
and call that when we need the functionality in both `processDir` used
by the `convert` command and the server functionality used by the
`serve` command

(3) hsblog-gen.cabal

We'll add 4 new packages as our dependencies:

- bytestring: provides a string of bytes data type (two variants, actually)
- http-types: provides types for talking about the http protocol
- wai: provides an API for defining web servers
- warp: provides a fast and lightweight web server for WAI applications

There are many choices we could've made regarding which web library to
use, in general I would've reached for `scotty` instead of wai, which
is more featureful and easier to use, but scotty depends on a lot more
packages than wai, and for our usecase wai provides everything we need.

(4) New file: src/Server.hs

In this new source file we'll define a wai `Application`. An
`Application` is just an alias for a function that takes an http
`Request` and a function that, when called with an http `Response`,
returns an execution plan that when executed will send that response
to the client and return whether it succeeded or not.

```
type Application =
  Request -> (Response -> IO ResponseReceived) -> IO ResponseReceived
```

Note that the cool thing here is that we are defining an abstract
application that will later call the function it received which we
will call `response`.

This gives the *caller* the ability to provide us with different
`Response -> IO ResponseReceived` function depending on the
implementation.

This could be a function that supports HTTP 1.0, HTTP 2.0, TLS, Or one
that mocks everything and doesn't actually do any networking at all!

In our case, we will use the function `run` from the `warp` package in
app/blog_gen.hs.

--

So the first thing we do at each request is to try and fetch and
convert all of the files in the input directory to html, as well as
create the index file.

This is an IO action that can fail, for example if the directory does
not exists. So instead of crashing the program, we'd like to handle
the exception gracefully and then decide what to do. We'll encode the
possible failure using `Either` which we've seen before, which
provides us with two alternative types with the constructors `Left`
and `Right`.

We will use the function `catch` from the `Control.Exception` module
to catch the exception if one is thrown and return its textual
representation as `Left`.

Check out the Operator Glossary or Hoogle for operators or functions
you are not familiar with:
https://www.fpcomplete.com/haskell/tutorial/operators/

Next, we will pattern match on that and the path info we got from the
request. If fetching the files failed, we will respond with an
internal error, otherwise we'll decide what to do according to the
path.

The request data type we got contains all of the information we need
to understand what the user requested. You can see the various things
we can get in the module documentation:

https://hackage.haskell.org/package/wai-3.2.2.1/docs/Network-Wai.html#g:3

In our case, we are interested in the request path info, which is the
address the user wants to reach (without the domain itself).
`pathInfo` formats it nicely for us in a list of `Text` elements
(splitted on the "/"), and without the query parameters. So if the
user entered just the domain, the result would be the empty list. If
the user tried accessing a page such as `blog-post.html`, the result
would be ["blog-post.html"]. We can decide with which page to respond
according to that path.

Our http responses are encoded as a status, header and body.

The status says what happened. Did the response succeed (status 200)?
Did the server software failed internally (status 500)? Maybe the page
does not exist (status 404)?

The header can contain many things, for our use case we will report
how the body should be read. Is it a music file, a javascript file,
plain text or html file? This is defined in the "Content-type" header.

Then we return the body, which is a bytestring, in our case encoded as
UTF8 of plain text html. Just like the ones we've been writing to files.

---

This is basically it. With that we've defined a few functions that
takes a `Request` and creates a `Response`, pass it to `respond`, and
we got a web server for our blog!